LayoutTests/imported/w3c/web-platform-tests/webmessaging/event.origin.sub-expected.txt

Blocked access to external URL http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/webmessaging/support/ChildWindowPostMessage.htm
CONSOLE MESSAGE: Unable to post message to http://xn--n8j6ds53lwwkrqhv28a.localhost:8800. Recipient has origin http://localhost:8800.


Harness Error (TIMEOUT), message = null

TIMEOUT Test Description: event.origin returns the origin of the message. Test timed out
